requirements.txt
Flask==2.3.3
line-bot-sdk==3.5.0
google-cloud-vision==3.4.4
Pillow==10.0.1
opencv-python==4.8.1.78
pytesseract==0.3.10
pandas==2.1.1
matplotlib==3.7.2
numpy==1.24.3
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
regex==2023.8.8
